+++ tools/Propulsion.Tool/Program.fs
@@ -207,9 +207,9 @@ let main argv =
                 let producer, disposeProducer =
                     match broker,topic with
                     | Some b,Some t ->
-                        let _ = Config.validateBrokerUri (Uri b)
-                        let cfg = KafkaProducerConfig.Create(appName, b, Confluent.Kafka.Acks.Leader, Confluent.Kafka.CompressionType.Lz4)
-                        let p = BatchedProducer.CreateWithConfigOverrides(log, cfg, t)
+                        let linger = FsKafka.Batching.BestEffortSerial (TimeSpan.FromMilliseconds 100.)
+                        let cfg = KafkaProducerConfig.Create(appName, b, Confluent.Kafka.Acks.Leader, linger, Confluent.Kafka.CompressionType.Lz4)
+                        let p = BatchedProducer.Create(log, cfg, t)
                         Some p, (p :> IDisposable).Dispose
                     | _ -> None, id
                 let projectBatch (log : ILogger) (ctx : IChangeFeedObserverContext) (docs : IReadOnlyList<Microsoft.Azure.Documents.Document>) = async {