+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ### Added
 ### Changed
+
+- `Kafka`: Targets [`FsKafka`/`FsKafka0` v `1.4.5`](https://github.com/jet/FsKafka/blob/master/CHANGELOG.md#1.4.5)
+
 ### Removed
 ### Fixed